pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jam-splitter"
version = "0.1.0"
description = "JamSplitter - Split your music into stems and generate lyrics"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
requires-python = ">=3.11"
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Topic :: Multimedia :: Sound/Audio :: Analysis",
    "Typing :: Typed",
    "Framework :: FastAPI",
    "Environment :: Web Environment"
]
dependencies = [
    "langchain>=0.0.300",
    "openai>=0.27.0",
    "openai-whisper>=20231117",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "typing-extensions>=4.5.0",
    "sqlalchemy-utils>=0.39.0",
    "class-inheritance>=1.0.0",
    "yt-dlp>=2025.6.0",
    "rich>=13.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "spotipy>=2.23.0",
    "psycopg2-binary>=2.9.0",
    "numpy>=1.24.0",
    "ffmpeg-python>=0.2.0",
    "tqdm>=4.65.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.0.0",
    "aiohttp>=3.9.0",
    "websockets>=12.0.0",
    "pytube>=15.0.0",
    "python-json-logger>=2.0.7",
    "prometheus-client>=0.19.0",
    "pyyaml>=6.0.1",
    "structlog>=23.2.0",
    "python-logstash>=0.4.6",
    "starlette-exporter>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-console-scripts>=1.0.0",
    "pytest-cov>=3.0.0",
    "flake8>=5.0.0",
    "black>=24.4.0",
    "isort>=5.13.2",
    "mypy>=1.10.0",
    "bandit>=1.7.7",
    "pre-commit>=3.0.0",
    "types-requests",
    "types-python-dateutil",
    "ruff>=0.1.0",
]

[tool.black]
line-length = 88
target-version = ["py311"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
known_first_party = ["app"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
check_untyped_defs = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "I", "N", "W", "B", "C4", "D", "UP"]
ignore = ["E501", "D107"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["S101", "D"]

[tool.ruff.flake8-quotes]
docstring-quotes = "double"

[tool.ruff.isort]
known-first-party = ["app"]

[tool.ruff.mccabe]
max-complexity = 10

[tool.bandit]
targets = ["app"]
recursive = true
skips = ["B101"]  # Skip assert_used check for tests

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
python_classes = ["Test*"]
addopts = "-v --cov=app --cov-report=term-missing"
filterwarnings = [
    "error",
    # Add any specific warnings to ignore here
]

[tool.coverage.run]
source = ["app"]
omit = ["app/tests/*"]

[tool.coverage.report]
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:"
]
